refactor (most of) the querying and scanning into the snowflake package
