+++ pkg/resources/warehouse.go
@@ -3,46 +3,11 @@ package resources
 import (
 	"database/sql"
 	"strings"
-	"time"
 
 	"github.com/chanzuckerberg/terraform-provider-snowflake/pkg/snowflake"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 )
-
-// warehouse is a go representation of a grant that can be used in conjunction
-// with github.com/jmoiron/sqlx
-type warehouse struct {
-	Name            string    `db:"name"`
-	State           string    `db:"state"`
-	Type            string    `db:"type"`
-	Size            string    `db:"size"`
-	MinClusterCount int64     `db:"min_cluster_count"`
-	MaxClusterCount int64     `db:"max_cluster_count"`
-	StartedClusters int64     `db:"started_clusters"`
-	Running         int64     `db:"running"`
-	Queued          int64     `db:"queued"`
-	IsDefault       string    `db:"is_default"`
-	IsCurrent       string    `db:"is_current"`
-	AutoSuspend     int64     `db:"auto_suspend"`
-	AutoResume      bool      `db:"auto_resume"`
-	Available       string    `db:"available"`
-	Provisioning    string    `db:"provisioning"`
-	Quiescing       string    `db:"quiescing"`
-	Other           string    `db:"other"`
-	CreatedOn       time.Time `db:"created_on"`
-	ResumedOn       time.Time `db:"resumed_on"`
-	UpdatedOn       time.Time `db:"updated_on"`
-	Owner           string    `db:"owner"`
-	Comment         string    `db:"comment"`
-	ResourceMonitor string    `db:"resource_monitor"`
-	Actives         int64     `db:"actives"`
-	Pendings        int64     `db:"pendings"`
-	Failed          int64     `db:"failed"`
-	Suspended       int64     `db:"suspended"`
-	UUID            string    `db:"uuid"`
-	ScalingPolicy   string    `db:"scaling_policy"`
-}
 
 // warehouseCreateProperties are only available via the CREATE statement
 var warehouseCreateProperties = []string{"initially_suspended", "wait_for_provisioning", "statement_timeout_in_seconds"}
@@ -166,10 +131,7 @@ func ReadWarehouse(data *schema.ResourceData, meta interface{}) error {
 	stmt := snowflake.Warehouse(data.Id()).Show()
 
 	row := snowflake.QueryRow(db, stmt)
-
-	w := &warehouse{}
-
-	err := row.StructScan(w)
+	w, err := snowflake.ScanWarehouse(row)
 	if err != nil {
 		return err
 	}
